update to jupyter
