+++ nosebook/__init__.py
@@ -1,0 +1,4 @@
+# flake8: noqa
+from ._version import *
+from .case import NoseCellTestCase
+from .plugin import Nosebook
