+++ nosebook/_version.py
@@ -1,0 +1,2 @@
+__version_info__ = (1, 0, 0)
+__version__ = '.'.join(map(str, __version_info__))
