+++ nosebook/tests/test_scrub.py
@@ -1,12 +1,13 @@
 import nosebook
-import munch
 
 import unittest
 
+class FakeKernel(object):
+    iopub_channel = {}
 
 class TestNosebook(unittest.TestCase):
     def setUp(self):
-        self.cell = munch.Munch({
+        self.cell = {
             "cell_type": "code",
             "execution_count": 21,
             "metadata": {
@@ -37,11 +38,9 @@ class TestNosebook(unittest.TestCase):
                 "print(1234)\n",
                 "rnd"
             ]
-        })
+        }
 
-        self.kernel = munch.Munch({
-            "iopub_channel": {}
-        })
+        self.kernel = FakeKernel()
 
     def make_case(self, scrubs, expected):
         case = nosebook.NoseCellTestCase(
@@ -52,9 +51,9 @@ class TestNosebook(unittest.TestCase):
         )
 
         self.assertEquals(
-            list(case.scrubOutputs(self.cell.outputs)),
+            list(case.scrubOutputs(self.cell["outputs"])),
             expected,
-            list(case.scrubOutputs(self.cell.outputs)),
+            list(case.scrubOutputs(self.cell["outputs"])),
         )
 
     def test_scrub_dict(self):
