+++ nosebook/util/dump_canonical.py
@@ -1,0 +1,5 @@
+import json
+
+
+def dump_canonical(obj):
+    return json.dumps(obj, indent=2, sort_keys=True)
