+++ requirements.txt
@@ -1,4 +1,4 @@
 nose>=1.3.4
-IPython
+nbformat
 jsonschema
 pyzmq
