+++ setup.py
@@ -1,16 +1,6 @@
 import os
 from setuptools import setup
 
-# you'd add this, too, for `python setup.py test` integration
-from setuptools.command.test import test as TestCommand
-
-
-class NosebookTestCommand(TestCommand):
-    def run_tests(self):
-        # Run nose ensuring that argv simulates running nosetests directly
-        import nose
-        nose.run_exit(argv=['nosetests', '-c', './.noserc'])
-
 
 def read(fname):
     """
@@ -27,9 +17,9 @@ setup(
     version="0.4.0",
     author="Nicholas Bollweg",
     author_email="nick.bollweg@gmail.com",
-    description="a nose plugin for IPython notebooks",
+    description="a nose plugin for Jupyter notebooks",
     license="BSD",
-    keywords="IPython nose",
+    keywords="Jupyter nose",
     url="http://github.com/bollwyvl/nosebook",
     py_modules=["nosebook"],
     long_description=read("README.rst"),
@@ -48,17 +38,16 @@ setup(
         "Topic :: Software Development :: Testing",
     ],
     setup_requires=[
-        "nose",
         "IPython",
-        "jsonschema",
-        "pyzmq"
+        "jupyter_client",
+        "nbformat",
+        "nose"
     ],
     entry_points={
         "nose.plugins.0.10": [
-            "nosebook = nosebook:Nosebook",
+            "nosebook = nosebook.plugin:Nosebook",
             "subprocstreams = "
             "IPython.testing.iptest:SubprocessStreamCapturePlugin"
         ]
-    },
-    cmdclass={'test': NosebookTestCommand}
+    }
 )
